labfiles/rag-app/python/requirements.txt
python-dotenv
flask==3.0.0
openai==1.54.4
gunicorn==21.2.0
flask-cors==4.0.0
requests==2.31.0
